Move replaceAll and its functions to a new class
